Moving unit/utils/test_jinja.py to pytest.
