--- tests/integration/files/file/base/templates/hello_import
@@ -1,3 +1,0 @@
-{% from 'macro' import mymacro -%}
-{% from 'macro' import mymacro -%}
-{{ mymacro('Hey') ~ mymacro(a|default('a'), b|default('b')) }}
