--- tests/integration/files/file/base/templates/macro
@@ -1,4 +1,0 @@
-# macro
-{% macro mymacro(greeting, greetee='world') -%}
-{{ greeting ~ ' ' ~ greetee }} !
-{%- endmacro %}
