+++ tests/pytests/unit/utils/jinja/test_jinja_default_options.py
@@ -1,0 +1,84 @@
+"""
+Tests for salt.utils.jinja
+"""
+
+import os
+
+import pytest
+import salt.config
+import salt.loader
+
+# dateutils is needed so that the strftime jinja filter is loaded
+import salt.utils.dateutils  # pylint: disable=unused-import
+import salt.utils.files
+import salt.utils.json
+import salt.utils.stringutils
+import salt.utils.yaml
+from salt.utils.templates import render_jinja_tmpl
+
+BLINESEP = salt.utils.stringutils.to_bytes(os.linesep)
+
+
+@pytest.fixture
+def minion_opts(tmpdir):
+    _opts = salt.config.DEFAULT_MINION_OPTS.copy()
+    _opts.update(
+        {
+            "cachedir": tmpdir.join("jinja-template-cache").strpath,
+            "file_buffer_size": 1048576,
+            "file_client": "local",
+            "file_ignore_regex": None,
+            "file_ignore_glob": None,
+            "file_roots": {"test": [tmpdir.join("templates").strpath]},
+            "pillar_roots": {"test": [tmpdir.join("templates").strpath]},
+            "fileserver_backend": ["roots"],
+            "hash_type": "md5",
+            "extension_modules": os.path.join(
+                os.path.dirname(os.path.abspath(__file__)), "extmods"
+            ),
+            "jinja_env": {"line_comment_prefix": "##", "line_statement_prefix": "%"},
+        }
+    )
+    return _opts
+
+
+@pytest.fixture
+def local_salt():
+    return {
+        "myvar": "zero",
+        "mylist": [0, 1, 2, 3],
+    }
+
+
+def test_comment_prefix(minion_opts, local_salt):
+
+    template = """
+        %- set myvar = 'one'
+        ## ignored comment 1
+        {{- myvar -}}
+        {%- set myvar = 'two' %} ## ignored comment 2
+        {{- myvar }} ## ignored comment 3
+        %- if myvar == 'two':
+        %- set myvar = 'three'
+        %- endif
+        {{- myvar -}}
+        """
+    rendered = render_jinja_tmpl(
+        template, dict(opts=minion_opts, saltenv="test", salt=local_salt)
+    )
+    assert rendered == "onetwothree"
+
+
+def test_statement_prefix(minion_opts, local_salt):
+
+    template = """
+        {%- set mylist = ['1', '2', '3'] %}
+        %- set mylist = ['one', 'two', 'three']
+        %- for item in mylist:
+        {{- item }}
+        %- endfor
+        """
+    rendered = render_jinja_tmpl(
+        template, dict(opts=minion_opts, saltenv="test", salt=local_salt)
+    )
+    assert rendered == "onetwothree"
